[Refactor]#2 data package 생성. 패키지 정리
- Clean Architecture로 리팩토링함에 따라 presentation 패키지 생성 후 파일 이동
- Req, Res 약어로 설정된 Remote Request, Response 파일들의 네이밍 변경
- Remote request, response, service 파일 분리
