+++ app/src/main/java/com/gallery/kakaogallery/data/remote/response/ImageSearchResponse.kt
@@ -1,42 +1,12 @@
-package com.gallery.kakaogallery.model
+package com.gallery.kakaogallery.data.remote.response
 
-import android.util.Log
-import com.gallery.kakaogallery.ApiAddressConstant
+import com.gallery.kakaogallery.model.QuerySearchModel
 import com.google.gson.annotations.SerializedName
-import io.reactivex.rxjava3.core.Flowable
-import retrofit2.http.GET
-import retrofit2.http.Query
-import java.lang.Exception
-import java.text.SimpleDateFormat
-import java.util.*
-import kotlin.collections.ArrayList
 
-interface ImageSearchService {
-    @GET(ApiAddressConstant.Image.ImageSearch)
-    fun requestSearchImage(
-        @Query("query") query : String,
-        @Query("sort") sort : String,
-        @Query("page") page : Int, // 결과 페이지 번호, 1~50 사이의 값, 기본 값 1
-        @Query("size") pageSize : Int //한 페이지에 보여질 문서 수, 1~80 사이의 값, 기본 값 80
-    ) : Flowable<ImageSearchResModel>
-}
 
-//https://developers.kakao.com/docs/latest/ko/daum-search/dev-guide#search-image
-class ImageSearchReqModel internal constructor(
-    val query : String,
-    val sort : SortType,
-    val page : Int, // 결과 페이지 번호, 1~50 사이의 값, 기본 값 1
-    val pageSize : Int // 한 페이지에 보여질 문서 수, 1~50 사이의 값, 기본 값 10
-){
-    enum class SortType(val key : String) {
-        Accuracy("accuracy"), // 정확도순
-        Recency("recency") // 최신순
-    }
-}
-
-class ImageSearchResModel {
+class ImageSearchResponse {
     @SerializedName("meta")
-    var imageSearchMetaData : ImageSearchMetaModel? = null
+    var imageSearchMetaData : ImageSearchMetaEntity? = null
     @SerializedName("documents")
     var imageSearchResList : ArrayList<ImageSearchModel> ? = null
 
@@ -45,7 +15,7 @@ class ImageSearchResModel {
     }
 }
 
-class ImageSearchMetaModel internal constructor(
+class ImageSearchMetaEntity internal constructor(
     @SerializedName("totle_count")val totalCount : Int, // 검색된 문서 수
     @SerializedName("pageable_count")val pageableCount : Int, // total_count 중 노출 가능 문서 수
     @SerializedName("is_end")val isEnd : Boolean // 현재 페이지가 마지막 페이지인지 여부, 값이 false면 page를 증가시켜 다음 페이지를 요청할 수 있음
