+++ app/src/main/java/com/gallery/kakaogallery/model/QuerySearchModel.kt
@@ -1,7 +1,5 @@
 package com.gallery.kakaogallery.model
 
-import android.util.Log
-import java.lang.Exception
 import java.text.SimpleDateFormat
 import java.util.*
