+++ app/src/main/java/com/gallery/kakaogallery/util/ApiUtil.kt
@@ -1,7 +1,12 @@
 package com.gallery.kakaogallery.util
 
 import android.util.Log
-import com.gallery.kakaogallery.model.*
+import com.gallery.kakaogallery.data.remote.request.ImageSearchRequest
+import com.gallery.kakaogallery.data.remote.request.VideoSearchRequest
+import com.gallery.kakaogallery.data.remote.response.ImageSearchResponse
+import com.gallery.kakaogallery.data.remote.response.VideoSearchResponse
+import com.gallery.kakaogallery.data.remote.service.ImageSearchService
+import com.gallery.kakaogallery.data.remote.service.VideoSearchService
 import com.gallery.kakaogallery.presentation.application.KakaoGalleryApplication
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import io.reactivex.rxjava3.core.Flowable
@@ -11,7 +16,7 @@ object ApiUtil : BaseUtil() {
     private val mRetrofit = KakaoGalleryApplication.mRetrofit
 
     object Image {
-        fun requestSearchImage(reqModel : ImageSearchReqModel) : Flowable<ImageSearchResModel>{
+        fun requestSearchImage(reqModel : ImageSearchRequest) : Flowable<ImageSearchResponse>{
             return mRetrofit.getService(ImageSearchService::class.java).run {
                 Log.d("TAG", "api request : ${Thread.currentThread().name}")
                 this.requestSearchImage(
@@ -26,7 +31,7 @@ object ApiUtil : BaseUtil() {
     }
 
     object Video {
-        fun requestSearchVideo(reqModel : VideoSearchReqModel) : Flowable<VideoSearchResModel>{
+        fun requestSearchVideo(reqModel : VideoSearchRequest) : Flowable<VideoSearchResponse>{
             return mRetrofit.getService(VideoSearchService::class.java).run {
                 this.requestSearchVideo(
                     reqModel.query,
